Remove cruft
