+++ crates/bindings-typescript/src/server/index.ts
@@ -1,2 +1,5 @@
 export * from './type_builders';
 export * from './type_util';
+export * from './schema';
+
+import './rt';
