+++ crates/bindings-typescript/src/server/type_builders.ts
@@ -3,7 +3,6 @@ import {
   ConnectionId,
   Identity,
   ScheduleAt,
-  SumTypeVariant,
   TimeDuration,
   Timestamp,
   type AlgebraicTypeVariants,
